File upload functionality
Now it is possible to upload files
Renamed some folders
Added new modules needed for visualizations
Edited the home page so that to test the functionality

Make sure to look at the changes and then how they conflict with your files in the local repository.

If there are any questions about this commit or if there is any conflict with your current local cloned repository and you don't know what to do, write me on Discord

If there isn't any conflict, be sure to pull first the files such that you update your local repository with the changes, to later commit your own changes.

If any of this does not make sense, feel free to write to me on Discord