Fixed issue with centering eye cloud after turning visualizations on/off
Fixed an issue where the eye cloud did not center properly anymore after turning visualizations on or off by redrawing the visualization upon a change of the box size. This could impact performance, so let me know if that's a major issue.

Furthermore, fixed an inaccuracy in the show info functionality and adjusted the eye cloud settings help message.